Testing Arduino disconnect patch
This changes the way the program detects whether the Arduino has
rebooted spontaneously. It sends out a test 'packet' to the
Arduino every second. When operating normally, the Arduino will
ignore this packet. But, if the Arduino has rebooted, the Arduino
will send back an 'invalid mode packet'. The program tries to
detect this and automatically attempts to restore the previous
operation mode.